cda rendering on view
