+++ package.json
@@ -23,7 +23,8 @@
     "axios": "^0.21.1",
     "bootstrap": "^4.6.0",
     "ejs": "^3.1.6",
-    "express": "^4.17.1"
+    "express": "^4.17.1",
+    "xml-formatter": "^2.4.0"
   },
   "devDependencies": {
     "@types/express": "^4.17.11",
