+++ src/index.ts
@@ -9,7 +9,7 @@ app.use(express.json());
 app.use(express.urlencoded({
     extended: true
 }));
-app.use(express.static('public'))
+app.use(express.static(path.join(__dirname, 'public')));
 
 // Configure Express to use EJS
 app.set("views", path.join(__dirname, "views"));
