+++ tsconfig.json
@@ -15,6 +15,6 @@
         }
     },
     "include": [
-        "src/**/*"
+        "src"
     ]
 }
